Add files via upload
NumPy is a Python library that provides support for large, multi-dimensional arrays and matrices, along with a large collection of high-level mathematical functions to operate on these arrays. 
NumPy is used in Python because it provides an array object that is faster and more efficient than traditional Python lists 1. The array object in NumPy is called Ndarray, and it provides a lot of supporting functions that make working with Ndarray very easy.

Arrays are very frequently used in data science, where speed and resources are very important 1. NumPy provides an array object that is up to 50x faster than traditional Python lists. NumPy also includes a wide range of mathematical functions, such as linear algebra, Fourier transforms, and random number generation, which can be applied to arrays.