Add files via upload
In this practice, the development of classifiers begins with selecting a classification dataset from the UCI Machine Learning Repository: https://archive.ics.uci.edu/. For simplicity and due to the lack of restrictions, the "Iris" dataset will be used in this work. This classic dataset, dating back to Fisher's 1936 paper, is a small and well-known dataset commonly used to evaluate classification methods. Given its widespread use and familiarity, it offers a valuable learning experience.